Took 47 seconds

+++ danielutils/retry_executor/backoff_strategy.py
@@ -6,10 +6,10 @@ class BackOffStrategy:
     A class to create a common abstraction for backoff strategies
     """
 
-    def __init__(self, supp: Supplier[int]) -> None:
+    def __init__(self, supp: Supplier[float]) -> None:
         self._supp = supp
 
-    def get_backoff(self) -> int:
+    def get_backoff(self) -> float:
         """
 
         :return: amount of milliseconds to sleep
